refactor: consolidate scanning logic into AST-based scanner and streamline directive and resolver scanning
